refactor: move navigation classes to model package and update imports
