+++ core/testing/di/src/jvmMain/kotlin/com/thomaskioko/tvmaniac/testing/di/TestJvmGraph.kt
@@ -2,7 +2,7 @@ package com.thomaskioko.tvmaniac.testing.di
 
 import com.thomaskioko.tvmaniac.datastore.api.DatastoreRepository
 import com.thomaskioko.tvmaniac.navigation.RootPresenter
-import com.thomaskioko.tvmaniac.navigation.ScreenGraph
+import com.thomaskioko.tvmaniac.navigation.di.ScreenGraph
 import com.thomaskioko.tvmaniac.traktauth.api.TraktAuthManager
 import dev.zacsweers.metro.AppScope
 import dev.zacsweers.metro.DependencyGraph
