+++ navigation/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/navigation/model/EpisodeSheetChild.kt
@@ -1,5 +1,6 @@
-package com.thomaskioko.tvmaniac.navigation
+package com.thomaskioko.tvmaniac.navigation.model
 
+import com.thomaskioko.tvmaniac.navigation.SheetChild
 import com.thomaskioko.tvmaniac.presentation.episodedetail.EpisodeDetailSheetPresenter
 
 public class EpisodeSheetChild(
