+++ navigation/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/navigation/model/NotificationPermissionState.kt
@@ -1,5 +1,8 @@
-package com.thomaskioko.tvmaniac.navigation
+package com.thomaskioko.tvmaniac.navigation.model
 
+import kotlinx.serialization.Serializable
+
+@Serializable
 public data class NotificationPermissionState(
     val showRationale: Boolean = false,
     val requestPermission: Boolean = false,
