+++ navigation/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/navigation/model/ThemeState.kt
@@ -1,7 +1,9 @@
-package com.thomaskioko.tvmaniac.navigation
+package com.thomaskioko.tvmaniac.navigation.model
 
 import com.thomaskioko.tvmaniac.datastore.api.AppTheme
+import kotlinx.serialization.Serializable
 
+@Serializable
 public data class ThemeState(
     val isFetching: Boolean = true,
     val appTheme: AppTheme = AppTheme.SYSTEM_THEME,
