+++ navigation/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/navigation/DefaultRootNavigator.kt
@@ -7,6 +7,7 @@ import com.arkivanov.decompose.router.stack.popTo
 import com.arkivanov.decompose.router.stack.pushNew
 import com.arkivanov.decompose.router.stack.pushToFront
 import com.thomaskioko.tvmaniac.core.base.ActivityScope
+import com.thomaskioko.tvmaniac.navigation.model.RootDestinationConfig
 import dev.zacsweers.metro.ContributesBinding
 import dev.zacsweers.metro.Inject
 import dev.zacsweers.metro.SingleIn
