+++ navigation/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/navigation/DefaultRootPresenter.kt
@@ -19,6 +19,13 @@ import com.thomaskioko.tvmaniac.core.view.collectStatus
 import com.thomaskioko.tvmaniac.datastore.api.DatastoreRepository
 import com.thomaskioko.tvmaniac.domain.logout.LogoutInteractor
 import com.thomaskioko.tvmaniac.domain.user.UpdateUserProfileData
+import com.thomaskioko.tvmaniac.navigation.di.ScreenGraph
+import com.thomaskioko.tvmaniac.navigation.model.DeepLinkDestination
+import com.thomaskioko.tvmaniac.navigation.model.EpisodeSheetChild
+import com.thomaskioko.tvmaniac.navigation.model.EpisodeSheetConfig
+import com.thomaskioko.tvmaniac.navigation.model.NotificationPermissionState
+import com.thomaskioko.tvmaniac.navigation.model.RootDestinationConfig
+import com.thomaskioko.tvmaniac.navigation.model.ThemeState
 import com.thomaskioko.tvmaniac.traktauth.api.AuthError
 import com.thomaskioko.tvmaniac.traktauth.api.TokenRefreshResult
 import com.thomaskioko.tvmaniac.traktauth.api.TraktAuthRepository
