+++ navigation/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/navigation/controllers/DefaultEpisodeSheetController.kt
@@ -5,8 +5,8 @@ import com.arkivanov.decompose.router.slot.activate
 import com.arkivanov.decompose.router.slot.dismiss
 import com.thomaskioko.nav.model.ScreenSource
 import com.thomaskioko.tvmaniac.core.base.ActivityScope
-import com.thomaskioko.tvmaniac.navigation.EpisodeSheetConfig
 import com.thomaskioko.tvmaniac.navigation.EpisodeSheetController
+import com.thomaskioko.tvmaniac.navigation.model.EpisodeSheetConfig
 import dev.zacsweers.metro.ContributesBinding
 import dev.zacsweers.metro.SingleIn
