+++ navigation/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/navigation/navigators/DefaultDiscoverNavigator.kt
@@ -6,8 +6,8 @@ import com.thomaskioko.nav.model.ShowDetailsParam
 import com.thomaskioko.tvmaniac.core.base.ActivityScope
 import com.thomaskioko.tvmaniac.discover.presenter.DiscoverNavigator
 import com.thomaskioko.tvmaniac.navigation.EpisodeSheetController
-import com.thomaskioko.tvmaniac.navigation.RootDestinationConfig
 import com.thomaskioko.tvmaniac.navigation.RootNavigator
+import com.thomaskioko.tvmaniac.navigation.model.RootDestinationConfig
 import com.thomaskioko.tvmaniac.presenter.home.HomeTabController
 import dev.zacsweers.metro.ContributesBinding
