+++ navigation/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/navigation/navigators/DefaultLibraryNavigator.kt
@@ -2,8 +2,8 @@ package com.thomaskioko.tvmaniac.navigation.navigators
 
 import com.thomaskioko.nav.model.ShowDetailsParam
 import com.thomaskioko.tvmaniac.core.base.ActivityScope
-import com.thomaskioko.tvmaniac.navigation.RootDestinationConfig
 import com.thomaskioko.tvmaniac.navigation.RootNavigator
+import com.thomaskioko.tvmaniac.navigation.model.RootDestinationConfig
 import com.thomaskioko.tvmaniac.presentation.library.LibraryNavigator
 import dev.zacsweers.metro.ContributesBinding
