+++ navigation/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/navigation/navigators/DefaultMoreShowsNavigator.kt
@@ -3,8 +3,8 @@ package com.thomaskioko.tvmaniac.navigation.navigators
 import com.thomaskioko.nav.model.ShowDetailsParam
 import com.thomaskioko.tvmaniac.core.base.ActivityScope
 import com.thomaskioko.tvmaniac.moreshows.presentation.MoreShowsNavigator
-import com.thomaskioko.tvmaniac.navigation.RootDestinationConfig
 import com.thomaskioko.tvmaniac.navigation.RootNavigator
+import com.thomaskioko.tvmaniac.navigation.model.RootDestinationConfig
 import dev.zacsweers.metro.ContributesBinding
 
 @ContributesBinding(ActivityScope::class)
