+++ navigation/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/navigation/navigators/DefaultSettingsNavigator.kt
@@ -1,8 +1,8 @@
 package com.thomaskioko.tvmaniac.navigation.navigators
 
 import com.thomaskioko.tvmaniac.core.base.ActivityScope
-import com.thomaskioko.tvmaniac.navigation.RootDestinationConfig
 import com.thomaskioko.tvmaniac.navigation.RootNavigator
+import com.thomaskioko.tvmaniac.navigation.model.RootDestinationConfig
 import com.thomaskioko.tvmaniac.settings.presenter.SettingsNavigator
 import dev.zacsweers.metro.ContributesBinding
