+++ navigation/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/navigation/navigators/DefaultUpNextNavigator.kt
@@ -5,8 +5,8 @@ import com.thomaskioko.nav.model.SeasonDetailsUiParam
 import com.thomaskioko.nav.model.ShowDetailsParam
 import com.thomaskioko.tvmaniac.core.base.ActivityScope
 import com.thomaskioko.tvmaniac.navigation.EpisodeSheetController
-import com.thomaskioko.tvmaniac.navigation.RootDestinationConfig
 import com.thomaskioko.tvmaniac.navigation.RootNavigator
+import com.thomaskioko.tvmaniac.navigation.model.RootDestinationConfig
 import com.thomaskioko.tvmaniac.presentation.upnext.UpNextNavigator
 import dev.zacsweers.metro.ContributesBinding
